+++ libinsitu/qc/graphs/main_layout.py
@@ -4,7 +4,7 @@ import numpy as np
 from matplotlib import pyplot as plt
 from matplotlib.gridspec import GridSpec, GridSpecFromSubplotSpec
 
-from libinsitu import info
+from libinsitu.log import info
 from libinsitu.qc.graphs.base import BaseGraphs, MC_CLEAR_COLOR, Text, GraphId, INDIVIDUAL_PLOTS, individual_graph, sub_grid
 
 STANDALONE_FONT_SIZE = 12